_Melhore código e implemente IA para reconhecer plankton_/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pillow>=10.0.0
opencv-python>=4.0.0
scikit-learn>=1.0.0


